+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added master volume control (plus/minus keys)
 - Added min/max altitude for car
+- Added chime sound for autopilot engage/disengage
 
 # 1.0.3